Move mergeLocalStorage to helpers.js
